add hygen template
