+++ _templates/game/new/containerDescriptionIndex.ts.ejs.t
@@ -1,0 +1,4 @@
+---
+to: src/containers/<%= name %>/Description/index.ts
+---
+export { default } from './Description';
