+++ _templates/game/new/containerHistoryTableIndex.ts.ejs.t
@@ -1,0 +1,4 @@
+---
+to: src/containers/<%= name %>/HistoryTable/index.ts
+---
+export { default } from './HistoryTable';
