+++ _templates/game/new/containerNewGame.test.tsx.ejs.t
@@ -1,13 +1,19 @@
+---
+to: src/containers/<%= name %>/NewGame/NewGame.test.tsx
+---
 import { ChakraProvider } from '@chakra-ui/react';
 import { render } from '@testing-library/react';
-import HistoryTable from './HistoryTable';
+import NewGame from './NewGame';
 import { createMatchMedia } from '@/lib/TestUtils/MatchMediaMock';
+
+jest.mock('next/router', () => ({ useRouter: jest.fn().mockReturnValue({ locale: 'en' }) }));
 
 test('should rendering', () => {
   window.matchMedia = createMatchMedia(1100);
+  const onNewGameMock = jest.fn();
   const { container } = render(
     <ChakraProvider>
-      <HistoryTable history={[]} />
+      <NewGame onNewGame={onNewGameMock} />
     </ChakraProvider>,
   );
   expect(container).toMatchSnapshot();
