+++ _templates/game/new/containerNewGameIndex.ts.ejs.t
@@ -1,0 +1,4 @@
+---
+to: src/containers/<%= name %>/NewGame/index.ts
+---
+export { default } from './NewGame';
