+++ _templates/game/new/game.ts.ejs.t
@@ -1,0 +1,58 @@
+---
+to: src/lib/<%= name %>Game/<%= name %>Game.ts
+---
+import Player from '../Player/Player';
+
+class <%= name %>Game {
+  private player: Player = new Player('Player1');
+  private roundScore: point[] = [];
+
+  resumeGame(progress: { round: point[]; score: point[][] }) {
+    for (const round of progress.score) {
+      this.player.roundScore(round);
+    }
+    this.roundScore = progress.round;
+  }
+  getRound() {
+    return this.player.getScore().length + 1;
+  }
+  getCurrentTarget() {
+    return 0;
+  }
+  addScore(score: point) {
+    if (this.roundScore.length >= 3) return;
+    this.roundScore.push(score);
+  }
+  removeScore() {
+    this.roundScore = [];
+  }
+  getScore() {
+    return this.player.getScore();
+  }
+  getRoundScore() {
+    return this.roundScore;
+  }
+  roundChange() {
+    if (this.roundScore.length > 3) return;
+    this.player.roundScore(this.roundScore);
+    this.roundScore = [];
+  }
+  getTotalScore() {
+    return 0;
+  }
+  isFinish() {
+    return false;
+  }
+  getProgressJson() {
+    return { round: this.roundScore, score: this.player.getScore() };
+  }
+  getGameResult(): <%= name %>Result {
+    return {
+      result: 0,
+      scores: [...this.player.getScore(), this.roundScore],
+      playedAt: new Date().toJSON(),
+    };
+  }
+}
+
+export default <%= name %>Game;
