+++ _templates/game/new/gameResultModel.ts.ejs.t
@@ -1,0 +1,10 @@
+---
+to: src/db/<%= name %>ResultModel.ts
+---
+import { ResultModel } from './ResultModel';
+
+export interface <%= name %>ResultModel extends ResultModel {
+  result: number;
+  scores: point[][];
+  playedAt: string;
+}
