+++ src/db/CricketMarkUpResultModel.ts
@@ -1,0 +1,14 @@
+import { ResultModel } from './ResultModel';
+
+export interface CricketMarkUpResultModel extends ResultModel {
+  targetCount: number;
+  result: number;
+  scores: CricketMarkUpScoreModel[];
+  playedAt: string;
+}
+
+export interface CricketMarkUpScoreModel {
+  number: number;
+  count: number;
+  total: number;
+}
