+++ src/db/DoubleTroubleResultModel.ts
@@ -1,0 +1,7 @@
+import { ResultModel } from './ResultModel';
+
+export interface DoubleTroubleResultModel extends ResultModel {
+  result: number;
+  scores: point[][];
+  playedAt: string;
+}
