+++ src/db/EaglesEyeResultModel.ts
@@ -1,0 +1,7 @@
+import { ResultModel } from './ResultModel';
+
+export interface EaglesEyeResultModel extends ResultModel {
+  result: number;
+  scores: point[][];
+  playedAt: string;
+}
