+++ src/db/ResultModel.ts
@@ -1,0 +1,3 @@
+export interface ResultModel {
+  id?: number;
+}
