+++ src/db/Sweet16ResultModel.ts
@@ -1,0 +1,7 @@
+import { ResultModel } from './ResultModel';
+
+export interface Sweet16ResultModel extends ResultModel {
+  result: number;
+  scores: point[][];
+  playedAt: string;
+}
