+++ src/db/TopsAndTensResultModel.ts
@@ -1,0 +1,7 @@
+import { ResultModel } from './ResultModel';
+
+export interface TopsAndTensResultModel extends ResultModel {
+  result: number;
+  scores: point[][];
+  playedAt: string;
+}
